backend/requirements.txt
fastapi
uvicorn
supabase
requests
python-dotenv
